requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==5.1.1
django-summernote==0.8.20.0
pillow==10.4.0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
webencodings==0.5.1
